Move `RebuildBootstrapper` command
