+++ wcfsetup/install/files/lib/system/package/PackageInstallationDispatcher.class.php
@@ -29,7 +29,6 @@ use wcf\system\form\element\MultipleSelectionFormElement;
 use wcf\system\form\element\TextInputFormElement;
 use wcf\system\form\FormDocument;
 use wcf\system\language\LanguageFactory;
-use wcf\system\package\command\RebuildBootstrapper;
 use wcf\system\package\plugin\IPackageInstallationPlugin;
 use wcf\system\registry\RegistryHandler;
 use wcf\system\request\RouteHandler;
@@ -217,7 +216,7 @@ class PackageInstallationDispatcher
 
             VersionTracker::getInstance()->createStorageTables();
 
-            $command = new RebuildBootstrapper();
+            $command = new \wcf\command\package\RebuildBootstrapper();
             $command();
 
             EventHandler::getInstance()->fire(new PackageListChanged());
