+++ wcfsetup/install/files/lib/system/package/PackageUninstallationDispatcher.class.php
@@ -9,7 +9,6 @@ use wcf\system\application\ApplicationHandler;
 use wcf\system\cache\builder\PackageCacheBuilder;
 use wcf\command\cache\ClearCache;
 use wcf\system\event\EventHandler;
-use wcf\system\package\command\RebuildBootstrapper;
 use wcf\system\package\plugin\IPackageInstallationPlugin;
 use wcf\system\setup\Uninstaller;
 use wcf\system\WCF;
@@ -78,7 +77,7 @@ class PackageUninstallationDispatcher extends PackageInstallationDispatcher
                     $step = $this->executePIP($nodeData);
 
                     if ($nodeData['pluginName'] == 'file') {
-                        $command = new RebuildBootstrapper();
+                        $command = new \wcf\command\package\RebuildBootstrapper();
                         $command();
                     }
                     break;
